move QualifierNav to components folder
