+++ frontend/src/pages/QualifierPage/components/QualifierNav.tsx
